Fix infinity loop
this commit is completely off pattern but as I copy the code a saw
various wrong executions that make it loop non stop, or just never end
the game.
This bug fix set off the infinity loop allowing player to input letters
until the game is over (with one extra input in the end for now).